pyproject.toml
[project]
name = "eat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "fastapi>=0.128.0",
    "langchain>=1.2.6",
    "langchain-openai>=0.3.0",
    "langgraph>=1.0.6",
    "langgraph-checkpoint-mysql>=2.0.17",
    "pymysql>=1.1.2",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.40.0",
]
